- dspEnvExpADSR: test, exp envelope - apps/env_test.cpp - plugin/host tweaks

+++ ax/axMain.h
@@ -2,15 +2,15 @@
  * This file is part of Axonlib.
  *
  * Axonlib is free software: you can redistribute it and/or modify
- * it under the terms of the Axonlib License, either version 1.0 
+ * it under the terms of the Axonlib License, either version 1.0
  * of the License, or (at your option) any later version.
  *
  * Axonlib is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  * See LICENSE_AX for more details.
- *  
- * You should have received a copy of the Axonlib License 
+ *
+ * You should have received a copy of the Axonlib License
  * If not, see <http://axonlib.googlecode.com/>.
  */
 
@@ -95,8 +95,8 @@
 
     int main(audioMasterCallback audioMaster)
     {
-      //AX_PLUGIN* plugin = new AX_PLUGIN(audioMaster,AX_NUMPROGS,AX_NUMPARAMS,AX_FLAGS);
-      axHost* host = new axHost((void*)audioMaster);
+      //axHost* host = new axHost((int)audioMaster);
+      axHost* host = new axHost((void*)(int)audioMaster);
       AX_PLUGIN* plugin = new AX_PLUGIN(host,AX_NUMPROGS,AX_NUMPARAMS,AX_FLAGS);
       if (!plugin) return 0;
       return (int)plugin->getAeffect();
@@ -122,7 +122,7 @@
 
     AEffect* main(audioMasterCallback audioMaster)
     {
-      //AX_PLUGIN* plugin = new AX_PLUGIN(audioMaster,AX_NUMPROGS,AX_NUMPARAMS,AX_FLAGS);
+      //axHost* host = new axHost((int)audioMaster);
       axHost* host = new axHost((void*)(int)audioMaster);
       AX_PLUGIN* plugin = new AX_PLUGIN(host,AX_NUMPROGS,AX_NUMPARAMS,AX_FLAGS);
       if (!plugin) return 0;
